scripts/requirements.txt
cwl_utils
miniwdl == 1.2.1

# major API change after 0.15.0
ruamel.yaml >= 0.15.0

pathlib == 1.0.1

google-cloud-storage
